Добавлен модуль отрисовки карточки. Проверка задания (7.10)
