+++ index.html
@@ -231,5 +231,6 @@
     </div>
   </template>
   <script src="js/main.js" type="module"></script>
+  <script src="js/generate-card.js" type="module"></script>
 </body>
 </html>
